Speed up specs by using shared_let.
